Add invert_permutation utility
